Made some huge updates - changed Home to Dashboard, put each widget into a separate file, removed the toolbar and replaced the fullscreen menu with a floating action button that expands on tap, and Improved the layout of the Hotel page to compensate for the toolbar removal.
